Fix cut-away view elevation to support negative values (#26752)
Change gClipHeight from uint8_t to int16_t to allow negative elevation
values in the cut-away view. Previously, the clip height was clamped to
0-255, preventing the cut-away view from showing underground when using
feet or metric units. The spinner, slider, and paint clipping code now
support a range of -128 to 255.

+++ src/openrct2/interface/Screenshot.cpp
@@ -49,7 +49,7 @@ using namespace OpenRCT2::Drawing;
 
 extern CoordsXY gClipSelectionA;
 extern CoordsXY gClipSelectionB;
-extern uint8_t gClipHeight;
+extern int16_t gClipHeight;
 
 uint8_t gScreenshotCountdown = 0;
