+++ src/openrct2/paint/Paint.h
@@ -237,7 +237,7 @@ struct PaintSession : public PaintSessionCore
 extern PaintSession gPaintSession;
 
 // Globals for paint clipping
-extern uint8_t gClipHeight;
+extern int16_t gClipHeight;
 extern CoordsXY gClipSelectionA;
 extern CoordsXY gClipSelectionB;
